move abbyy spec to its own file; put ocr output in its own subfolder by druid
