+++ config/settings.yml
@@ -42,4 +42,3 @@ purl_url: 'https://purl-example.stanford.edu'
 preservation_catalog:
   url: 'https://preservation-catalog-stage-01.stanford.edu'
   token: 'mint-token-with-target-preservation-catalog-rake-generate-token'
-
