+++ lib/dor/text_extraction/abbyy.rb
@@ -4,13 +4,15 @@ module Dor
   module TextExtraction
     # Build Abbyy tickets for processing
     class Abbyy
+      attr_reader :filepaths, :druid
+
       def initialize(filepaths:, druid:)
         @filepaths = filepaths
         @druid = druid
       end
 
       def abbyy_ticket_filepath
-        File.join(Settings.sdr.abbyy_ticket_path, "#{@druid}.xml")
+        File.join(Settings.sdr.abbyy_ticket_path, "#{druid}.xml")
       end
 
       def image_abbyy
@@ -18,12 +20,16 @@ module Dor
       end
 
       def pdf_files?
-        File.extname(@filepaths.first).strip.downcase[1..] == 'pdf'
+        File.extname(filepaths.first).strip.downcase[1..] == 'pdf'
+      end
+
+      def output_file_path
+        File.join(Settings.sdr.abbyy_output_path, druid).encode(xml: :text)
       end
 
       def alto_output
         "<ExportFormat OutputFileFormat='ALTO' OutputFlowType='SharedFolder' FormatVersion='3_1' CoordinatesParticularity='Words' WriteWordConfidence='true'>
-          <OutputLocation>#{Settings.sdr.abbyy_output_path.encode(xml: :text)}</OutputLocation>
+          <OutputLocation>#{output_file_path}</OutputLocation>
           <FileExistsAction>Overwrite</FileExistsAction>
           <NamingRule>&lt;FileName&gt;.&lt;Ext&gt;</NamingRule>
         </ExportFormat>"
@@ -31,7 +37,7 @@ module Dor
 
       def text_output
         "<ExportFormat OutputFileFormat='Text' OutputFlowType='SharedFolder' EncodingType='UTF8'>
-          <OutputLocation>#{Settings.sdr.abbyy_output_path.encode(xml: :text)}</OutputLocation>
+          <OutputLocation>#{output_file_path}</OutputLocation>
           <FileExistsAction>Overwrite</FileExistsAction>
           <NamingRule>&lt;FileName&gt;.&lt;Ext&gt;</NamingRule>
         </ExportFormat>"
@@ -42,14 +48,14 @@ module Dor
                        PdfAComplianceMode='PDFAM_PdfA_2u' PdfVersion='Version17'
                        Scenario='MaxQuality' UseImprovedCompression='true' MRCMode='Normal'
                        PdfUACompatible='true'>
-          <OutputLocation>#{Settings.sdr.abbyy_output_path.encode(xml: :text)}</OutputLocation>
+          <OutputLocation>#{output_file_path}</OutputLocation>
           <FileExistsAction>Overwrite</FileExistsAction>
           <NamingRule>&lt;FileName&gt;.&lt;Ext&gt;</NamingRule>
         </ExportFormat>"
       end
 
       def input_filepaths_field
-        @filepaths.map { |filename| "<InputFile Name=#{filename.encode(xml: :attr)}/>" }.join("\n")
+        filepaths.map { |filename| "<InputFile Name=#{filename.encode(xml: :attr)}/>" }.join("\n")
       end
 
       def xml_ticket
