+++ spec/fixtures/ocr/druid:bb222cc3333_abbyy_ticket.xml
@@ -5,17 +5,17 @@
                 PdfAComplianceMode='PDFAM_PdfA_2u' PdfVersion='Version17'
                 Scenario='MaxQuality' UseImprovedCompression='true' MRCMode='Normal'
                 PdfUACompatible='true'>
-      <OutputLocation>/tmp</OutputLocation>
+      <OutputLocation>/tmp/druid:bb222cc3333</OutputLocation>
       <FileExistsAction>Overwrite</FileExistsAction>
       <NamingRule>&lt;FileName&gt;.&lt;Ext&gt;</NamingRule>
     </ExportFormat>
     <ExportFormat OutputFileFormat='ALTO' OutputFlowType='SharedFolder' FormatVersion='3_1' CoordinatesParticularity='Words' WriteWordConfidence='true'>
-      <OutputLocation>/tmp</OutputLocation>
+      <OutputLocation>/tmp/druid:bb222cc3333</OutputLocation>
       <FileExistsAction>Overwrite</FileExistsAction>
       <NamingRule>&lt;FileName&gt;.&lt;Ext&gt;</NamingRule>
     </ExportFormat>
     <ExportFormat OutputFileFormat='Text' OutputFlowType='SharedFolder' EncodingType='UTF8'>
-      <OutputLocation>/tmp</OutputLocation>
+      <OutputLocation>/tmp/druid:bb222cc3333</OutputLocation>
       <FileExistsAction>Overwrite</FileExistsAction>
       <NamingRule>&lt;FileName&gt;.&lt;Ext&gt;</NamingRule>
     </ExportFormat>
